Move init tests to `new.rs` file
