+++ crates/forge/tests/e2e/mod.rs
@@ -20,6 +20,7 @@ mod fork_warning;
 mod forking;
 mod fuzzing;
 mod io_operations;
+mod new;
 mod requirements;
 mod running;
 mod steps;
